# 测试全流程框架/10_api_test_接口测试/01_api_docs/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: CRM系统API文档
  description: |
    客户关系管理系统后端API接口文档

    ## 规范说明
    - 用户名：4-50字符，支持中文、英文、数字、下划线
    - 密码：6-32字符，支持英文、数字、特殊字符
    - 客户名称：2-100字符
    - 联系电话：11位手机号
    - 跟进内容：10-2000字符
    - 合同金额/预估金额：正数
host: localhost:3000
basePath: /api
schemes:
  - http
definitions:
  Error:
    type: object
    properties:
      error:
        type: string
        description: 错误信息
  User:
    type: object
    properties:
      id:
        type: integer
        description: 用户ID
      username:
        type: string
        description: 用户名（4-50字符，支持中文、英文、数字、下划线）
      name:
        type: string
        description: 姓名
      phone:
        type: string
        description: 手机号（11位）
      email:
        type: string
        description: 邮箱
      role:
        type: string
        description: 角色
  Customer:
    type: object
    properties:
      id:
        type: integer
        description: 客户ID
      name:
        type: string
        description: 客户名称（2-100字符）
      phone:
        type: string
        description: 联系电话（11位手机号，脱敏显示：176****3123）
      industry:
        type: string
        description: 所属行业
      level:
        type: string
        description: 客户等级
      tag:
        type: array
        items:
          type: string
        description: 客户标签（最多3个）
      next_follow_time:
        type: string
        format: date
        description: 下次跟进时间（YYYY-MM-DD）
      business_count:
        type: integer
        description: 关联商机数量
      remark:
        type: string
        description: 客户备注（最大1000字符）
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
...
paths:
  /users/login:
    post:
      summary: 用户登录
      description: |
        用户登录接口

        **业务规则**：
        - 用户名：4-50字符，支持中文、英文、数字、下划线
        - 密码：6-32字符，支持英文、数字、特殊字符
        - 连续登录失败5次，账号锁定10分钟
        - 登录成功创建会话，有效期2小时
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - username
              - password
            properties:
              username:
                type: string
                description: 用户名（4-50字符，支持中文、英文、数字、下划线）
                minLength: 4
                maxLength: 50
                example: "admin_zhangsan"
              password:
                type: string
                description: 密码（6-32字符）
                minLength: 6
                maxLength: 32
                example: "Pass1234!"
      responses:
        200:
          description: 登录成功
          examples:
            application/json:
              {
                "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoxLCJ1c2VybmFtZSI6ImFkbWluX3poYW5nc2FuIiwiZXhwIjoxNzEzMDAwMDAwfQ.abc123xyz789",
                "user": {
                  "id": 1,
                  "username": "admin_zhangsan",
                  "name": "张三",
                  "phone": "13800138000",
                  "email": "zhangsan@example.com",
                  "role": "admin"
                }
              }
          schema:
            type: object
            properties:
              token:
                type: string
                description: 会话token（有效期2小时）
              user:
                $ref: '#/definitions/User'
        400:
          description: 参数错误（用户名或密码长度不符合要求）
          examples:
            application/json:
              {
                "error": "用户名或密码长度不符合要求"
              }
          schema:
            $ref: '#/definitions/Error'
        401:
          description: 用户名或密码错误
          examples:
            application/json:
              {
                "error": "用户名或密码错误"
              }
          schema:
            $ref: '#/definitions/Error'
        423:
          description: 账号已锁定
          examples:
            application/json:
              {
                "error": "账号已锁定，请10分钟后再试"
              }
          schema:
            $ref: '#/definitions/Error'
...
  /users/info:
    get:
      summary: 获取当前用户信息
      description: 获取已登录用户的基本信息
      responses:
        200:
          description: 成功
          schema:
            $ref: '#/definitions/User'
        401:
          description: 未登录或会话已过期
          schema:
            $ref: '#/definitions/Error'
  /customers:
    get:
      summary: 获取客户列表
      description: |
        获取客户列表，支持搜索和筛选

        **筛选条件**：
        - industry: 行业
        - level: 客户等级
        - tag: 客户标签
        - created_at_start/end: 创建时间范围
      parameters:
        - name: search
          in: query
          type: string
          description: 搜索关键词（客户名称、电话、标签）
        - name: industry
          in: query
          type: string
          description: 行业筛选
        - name: level
          in: query
          type: string
          description: 客户等级筛选
        - name: page
          in: query
          type: integer
          default: 1
          description: 页码
        - name: page_size
          in: query
          type: integer
          default: 20
          description: 每页条数
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              total:
                type: integer
              list:
                type: array
                items:
                  $ref: '#/definitions/Customer'

    post:
      summary: 新增客户
      description: |
        创建新客户

        **字段规范**：
        - 客户名称：2-100字符，必填
        - 联系电话：11位手机号，必填
        - 客户标签：最多3个
        - 客户备注：最大1000字符
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - name
              - phone
            properties:
              name:
                type: string
                description: 客户名称（2-100字符）
                minLength: 2
                maxLength: 100
              phone:
                type: string
                description: 联系电话（11位手机号）
                pattern: '^1[3-9]\d{9}$'
              industry:
                type: string
                description: 所属行业
              level:
                type: string
                description: 客户等级
              tag:
                type: array
                items:
                  type: string
                maxItems: 3
                description: 客户标签（最多3个）
              remark:
                type: string
                description: 客户备注（最大1000字符）
                maxLength: 1000
      responses:
        201:
          description: 创建成功
          schema:
            $ref: '#/definitions/Customer'
        400:
          description: 参数错误
          schema:
            $ref: '#/definitions/Error'
  /customers/{id}:
    get:
      summary: 获取客户详情
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: 客户ID
      responses:
        200:
          description: 成功
          schema:
            $ref: '#/definitions/Customer'
        404:
          description: 客户不存在
          schema:
            $ref: '#/definitions/Error'

    put:
      summary: 更新客户信息
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: 客户ID
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                description: 客户名称（2-100字符）
                minLength: 2
                maxLength: 100
              phone:
                type: string
                description: 联系电话（11位手机号）
              industry:
                type: string
                description: 所属行业
              level:
                type: string
                description: 客户等级
              tag:
                type: array
                items:
                  type: string
                maxItems: 3
              remark:
                type: string
                description: 客户备注（最大1000字符）
                maxLength: 1000
      responses:
        200:
          description: 更新成功
          schema:
            $ref: '#/definitions/Customer'
        400:
          description: 参数错误
          schema:
            $ref: '#/definitions/Error'
        404:
          description: 客户不存在
          schema:
            $ref: '#/definitions/Error'

    delete:
      summary: 删除客户
      description: 删除客户（需二次确认）
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: 客户ID
      responses:
        200:
          description: 删除成功
        404:
          description: 客户不存在
          schema:
            $ref: '#/definitions/Error'
  /followups:
    get:
      summary: 获取跟进记录列表
      parameters:
        - name: customer_id
          in: query
          type: integer
          description: 客户ID筛选
        - name: search
          in: query
          type: string
          description: 搜索关键词（客户名称、跟进内容）
        - name: page
          in: query
          type: integer
          default: 1
        - name: page_size
          in: query
          type: integer
          default: 20
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              total:
                type: integer
              list:
                type: array
                items:
                  $ref: '#/definitions/FollowUp'

    post:
      summary: 新增跟进记录
      description: |
        添加跟进记录

        **字段规范**：
        - 关联客户：必填
        - 跟进方式：电话/拜访/微信/邮件/短信，必填
        - 跟进内容：10-2000字符，必填
        - 下次跟进时间：不能早于当前时间
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - customer_id
              - method
              - content
            properties:
              customer_id:
                type: integer
                description: 关联客户ID
              method:
                type: string
                enum: [电话, 拜访, 微信, 邮件, 短信]
                description: 跟进方式
              content:
                type: string
                description: 跟进内容（10-2000字符）
                minLength: 10
                maxLength: 2000
              follow_time:
                type: string
                format: date-time
                description: 跟进时间
              next_follow_time:
                type: string
                format: date
                description: 下次跟进时间（不能早于当前时间）
      responses:
        201:
          description: 创建成功
          schema:
            $ref: '#/definitions/FollowUp'
        400:
          description: 参数错误
          schema:
            $ref: '#/definitions/Error'
  /followups/{id}:
    put:
      summary: 编辑跟进记录
      description: 编辑已有的跟进记录
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: 跟进记录ID
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              method:
                type: string
                enum: [电话, 拜访, 微信, 邮件, 短信]
                description: 跟进方式
              content:
                type: string
                description: 跟进内容（10-2000字符）
                minLength: 10
                maxLength: 2000
              next_follow_time:
                type: string
                format: date
                description: 下次跟进时间（不能早于当前时间）
      responses:
        200:
          description: 更新成功
          schema:
            $ref: '#/definitions/FollowUp'
        400:
          description: 参数错误
          schema:
            $ref: '#/definitions/Error'
        404:
          description: 跟进记录不存在
          schema:
            $ref: '#/definitions/Error'
  /business:
    get:
      summary: 获取商机列表
      parameters:
        - name: customer_id
          in: query
          type: integer
          description: 客户ID筛选
        - name: stage
          in: query
          type: string
          description: 商机阶段筛选
        - name: search
          in: query
          type: string
          description: 搜索关键词
        - name: page
          in: query
          type: integer
          default: 1
        - name: page_size
          in: query
          type: integer
          default: 20
      responses:
        200:
          description: 成功
          schema:
            type: object
            properties:
              total:
                type: integer
              list:
                type: array
                items:
                  $ref: '#/definitions/Business'

    post:
      summary: 新增商机
      description: |
        创建新商机

        **字段规范**：
        - 商机名称：最大100字符，必填
        - 关联客户：必填
        - 预估金额：正数
        - 商机阶段：必填（需求调研/方案制定/商务谈判/合同签订/成交）
        - 预计成交日期：不能早于当前日期
        - 阶段只能向前推进，不能回退
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - name
              - customer_id
              - stage
            properties:
              customer_id:
                type: integer
                description: 关联客户ID
              name:
                type: string
                description: 商机名称（最大100字符）
                maxLength: 100
              amount:
                type: number
                format: float
                minimum: 0
                description: 预估金额（正数）
              stage:
                type: string
                enum: [需求调研, 方案制定, 商务谈判, 合同签订, 成交]
                description: 商机阶段
              probability:
                type: integer
                minimum: 0
                maximum: 100
                description: 赢单概率（0-100%）
              expected_date:
                type: string
                format: date
                description: 预计成交日期（不能早于当前日期）
              remark:
                type: string
                description: 备注
      responses:
        201:
          description: 创建成功
          schema:
            $ref: '#/definitions/Business'
        400:
          description: 参数错误
          schema:
            $ref: '#/definitions/Error'
  /business/{id}:
    get:
      summary: 获取商机详情
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: 商机ID
      responses:
        200:
          description: 成功
          schema:
            $ref: '#/definitions/Business'
        404:
          description: 商机不存在
          schema:
            $ref: '#/definitions/Error'

    put:
      summary: 更新商机
      description: |
        更新商机信息

        **业务规则**：
        - 商机阶段只能向前推进，不能回退
        - 例如：从"需求调研"可以推进到"方案制定"，但不能退回
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: 商机ID
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
                description: 商机名称（最大100字符）
                maxLength: 100
              amount:
                type: number
                format: float
                minimum: 0
                description: 预估金额（正数）
              stage:
                type: string
                enum: [需求调研, 方案制定, 商务谈判, 合同签订, 成交]
                description: 商机阶段（只能向前推进）
              probability:
                type: integer
                minimum: 0
                maximum: 100
                description: 赢单概率（0-100%）
              expected_date:
                type: string
                format: date
                description: 预计成交日期（不能早于当前日期）
              remark:
                type: string
                description: 备注
      responses:
        200:
          description: 更新成功
          schema:
            $ref: '#/definitions/Business'
        400:
          description: 参数错误或阶段不能回退
          schema:
            $ref: '#/definitions/Error'
        404:
          description: 商机不存在
          schema:
            $ref: '#/definitions/Error'

    delete:
      summary: 删除商机
      description: 删除商机（需二次确认）
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: 商机ID
      responses:
        200:
          description: 删除成功
        404:
          description: 商机不存在
          schema:
            $ref: '#/definitions/Error'
...
  /approvals/{id}:
    put:
      summary: 审批操作
      description: |
        审批操作（同意/拒绝）

        **操作规则**：
        - 同意：审批状态变更为"已通过"
        - 拒绝：审批状态变更为"已拒绝"
      parameters:
        - name: id
          in: path
          type: integer
          required: true
          description: 审批ID
        - name: body
          in: body
          required: true
          schema:
            type: object
            required:
              - status
            properties:
              status:
                type: string
                enum: [已通过, 已拒绝]
                description: 审批状态
              remark:
                type: string
                description: 审批备注
      responses:
        200:
          description: 审批成功
          schema:
            $ref: '#/definitions/Approval'
        400:
          description: 参数错误
          schema:
            $ref: '#/definitions/Error'
        404:
          description: 审批记录不存在
          schema:
            $ref: '#/definitions/Error'
...